chore(workers): core poker-hand-rank + tag casino hand-eval helpers (#2126 r2)
- Move PokerHand* rank constants + PokerHandNames from PokerPlayer.go (casino)
  to a new untagged core file poker_hand_rank.go: core hand_eval.go/kicker.go
  and PokerSquares (solo) depend on them, so they must be in every worker.
- Tag the casino-only hand-eval/equity helpers (badugi/deuce_to_seven/shortdeck
  hand-eval, equity_engine) casino — they reference casino types.
