+++ internal/domain/equity_engine.go
@@ -1,3 +1,5 @@
+//go:build !js || !wasm || casino
+
 package domain
 
 import "math/rand"
